requirements.txt
Flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
dnspython==2.6.1
lxml==5.2.1
